requirements.txt
boto3
argparse
pyyaml